Move flask globals in cornac.flask package
